temporarily remove unnecessary code
This appears to be used to compile the elisp sources to bytecode if they are
not already compiled. The problem is it messes up installs that already do the
bytecoding and have the elc files only. It fails badly also if it can't find the
makefile. Need to figure out why and what problem this solves.